Cap home dependency version below 5.11
Home released version 5.11 bumping the msrv to rust 1.81.  They justify
such an aggressive msrv policy by stating that users should not be using
the home crate at all, but instead should rely on the standard library
home_dir function.  That function has a bug that is not fixed until
1.85, which prevents it from being an option for lower rust versions.

To mitigate, just pin to an older home version for now.